pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "one-click-campaign"
version = "0.1.0"
description = "LLM Agent to convert natural language into Pega When Rules"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.5.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
pega-agent = "pega_agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
